Include a tip about git in the start guide
Missing files and folders in the new blog folder is an issue I encountered. When this issue arised this afternoon, I found the solutions on Stackoverflow, which asked me to install git and add environment variables for git. The issues were resolved. 

As an absolute beginner, I have not installed git on my new work laptop. I am a bit surprised that nothing about installing git is mentioned in the start guide. Although this issue may seem trivial for many, I still think it would be better to have a tip about it so that future beginners do not need to spend extra time on this issue.

The link to the solution of my issue is: https://stackoverflow.com/questions/59442185/in-windows-rails-new-does-not-create-all-the-files-and-folders-stops-at-r